fix edge

+++ module3/task5/sd/edge/edge.h
@@ -7,9 +7,9 @@
 
 
 struct Edge {
-    int from;
-    int to;
-    int weight;
+    double from;
+    double to;
+    double weight;
 };
